Completed intro section, added more config variables, added osu folder detection, etc
Completed the intro section, which features the artist name, song title, who played the song, accuracy, combo / max combo, score, miss count, and overall rank. Some customization-oriented configuration was added, such as control over compression, framerate, and resolution. Utilities (7zip, oppai and ffmpeg) were moved into their own folder, along with fonts. The user's osu folder is also now automatically set to %appdata%/Local/osu!, but if the folder is elsewhere is asks for the user to assign it.

+++ RecordTheBeat/RecordTheBeat/Data/HitObjects/TimingPoint.cs
@@ -2,7 +2,7 @@
 {
     public class TimingPoint
     {
-        public int Offset { get; set; }
+        public double Offset { get; set; }
         public double MillisPerBeat { get; set; }
         public int Meter { get; set; }
         public int SampleSet { get; set; }
